server/Chatbot/requirement.txt
asgiref==3.3.4
click==8.0.1
colorama==0.4.4
Django==3.2.3
djangorestframework==3.12.4
joblib==1.0.1
nltk==3.6.2
numpy==1.20.3
pandas==1.2.4
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
six==1.16.0
sqlparse==0.4.1
tqdm==4.61.0
